third_party/blink/web_tests/http/tests/origin_trials/webexposed/animationworklet-origin-trial-interfaces.html

<!DOCTYPE html>
<!-- Generate token with the command:
generate_token.py http://127.0.0.1:8000 AnimationWorklet --expire-timestamp=2000000000
-->
<title>AnimationWorklet - interfaces exposed by origin trial</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/origin-trials-helper.js"></script>
<script>
const token="AnRfiFGg50a1+vvDBVfbBNiiRFBBQDy+3oMlJw2pUNRb8tPyN2XX0kAJxiA0itCLJTM0DrsIfVYuv4ni0Lu6bQgAAABYeyJvcmlnaW4iOiAiaHR0cDovLzEyNy4wLjAuMTo4MDAwIiwgImZlYXR1cmUiOiAiQW5pbWF0aW9uV29ya2xldCIsICJleHBpcnkiOiAyMDAwMDAwMDAwfQ=="

interfaces_to_check = ['WorkletAnimation', 'ScrollTimeline'];

properties_to_check = {
  'WorkletAnimation': ['play', 'cancel', 'playState', 'timeline'],
  'ScrollTimeline': ['scrollSource', 'orientation', 'timeRange'],
};

// Check that the interfaces are not present without the token.
// Experimental web platform features are always enabled when this test is run
// normally, so do not check them in that case.
if (!self.internals.runtimeFlags.animationWorkletEnabled) {
  test(t => {
    OriginTrialsHelper.check_interfaces_missing(this, interfaces_to_check);
  }, 'AnimationWorklet related interfaces are not available without a token.');
}

OriginTrialsHelper.add_token(token);

test(t => {
  OriginTrialsHelper.check_interfaces_exist(this, interfaces_to_check);
}, 'AnimationWorklet related interfaces in Origin-Trial enabled document.');

test(t => {
  assert_own_property(CSS, 'animationWorklet');
  OriginTrialsHelper.check_properties_exist(this, properties_to_check);
}, 'AnimationWorklet related properties in Origin-Trial enabled document.');

</script>
